+++ MindBudget/Models/Schema/SchemaV2.swift
@@ -45,3 +45,29 @@ enum SchemaV3: VersionedSchema {
         ]
     }
 }
+
+enum SchemaV4: VersionedSchema {
+    static var versionIdentifier: Schema.Version {
+        Schema.Version(4, 0, 0)
+    }
+
+    static var models: [any PersistentModel.Type] {
+        [
+            Expense.self,
+            Income.self,
+            IncomeAllocation.self,
+            SavingsGoal.self,
+            RecurringFixedExpenseRule.self,
+            RecurringExpenseOccurrence.self,
+            BudgetPlan.self,
+            BudgetPlanSemantics.self,
+            CategoryBudget.self,
+            WishItem.self,
+            CoolingOffPlan.self,
+            SpendingInsight.self,
+            ReflectionLog.self,
+            Merchant.self,
+            ReminderEvent.self
+        ]
+    }
+}